+++ BPMobileMessaging/Classes/ConactCenterCommunication/Model/ContactCenterVersion.swift
@@ -1,0 +1,15 @@
+//
+//  ContactCenterVersion.swift
+//  BPMobileMessaging
+//
+//  Created by Alexander Lobastov on 4/12/21.
+//
+
+import Foundation
+
+/// Describes server version.
+/// - Tag: ContactCenterVersion
+public struct ContactCenterVersion {
+    /// Server version
+    public let serverVersion: String
+}